+++ entity/constant/constants.go
@@ -1,0 +1,13 @@
+package constant
+
+// 错误码定义
+const (
+	// 成功
+	CodeSuccess = 200
+	// 参数错误
+	CodeInvalidParams = 400
+	// 未找到记录
+	CodeNotFound = 404
+	// 内部服务器错误
+	CodeServerError = 500
+)
